// backend/src/integrations/odds-corp/types/update-event.ts
export type UpdateEvent = {
  bk_event_id: string;
  bk_event_native_id: string;
  bk_id: string;
  bk_name: string;
  direct_link: string;
  event_name: string;
  league_name: string;
  live_delay: null | string;
  meta: string;
  score: string;
  sport: string;
  sport_id: string;
  team1: string;
  team2: string;
};